Updated python implementation
During this implementation the XML tree was imported so we can write our video stream's detections data into an xml file. The user can give the object of interest to be tracked. A demo output can be found at the file session.xml and a demonstration video was added as well under the demonstration-videos folder.